dba_agent/requirements.txt
google-adk[database]==0.3.0
apscheduler==3.11.0
schedule==1.2.2